Rename navigation source resolver module
